pyproject.toml
[tool.poetry]
name = "contrast-api"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">3.10,<3.11"
Django = "^4.2"
django-configurations = "^2.5"
dj-database-url = "^2.1.0"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
django-extensions = "^3.2.3"
django-import-export = "^3.3.1"
django-storages = "^1.14"
Pillow = "^10.1.0"
django-cors-headers = "^4.3.0"
django-querycount = "^0.8.3"
gunicorn = "^21.2.0"
djangorestframework-simplejwt = "^5.3.0"
psycopg2 = "^2.9.9"
drf-nested-routers = "^0.93.4"
drf-spectacular = "^0.26.5"
boto3 = ">=1.26.109"
django-nested-inline = "^0.4.6"
django-modeladmin-reorder = {git = "https://github.com/bloodwithmilk25/django-modeladmin-reorder.git", rev = "master"}
django-spa = {git = "https://github.com/VBobCat/django-spa.git", rev = "master"}
django-countries = "^7.5.1"
numpy = "^1.26.1"
pandas = "^1.5.3"
whitenoise = "6.2.0"
django-admin-rangefilter = "^0.11.2"
django-anymail = "^10.2"

sentry-sdk = "^1.32.0"
django-two-factor-auth = { extras = ["phonenumbers"], version = "^1.15.5" }
django-simple-history = "^3.4.0"
[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
ruff = "^0.1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

# Never enforce `E501` (line length violations).
ignore = ["E501"]